fix: optimise htmlDiff file code
